[IMP] websale_stock: add email notification
This PR adds a message "Be notified when in stock" on a product when it is out
of stock. When clicking the message, a form appears where the user can enter
his email. If he is a logged user, the email associated to the partner of the
user is pre-filled in the form. Otherwise the user enters his email.
Afterwards, the user clicks on the button and the following things take place:

   1. A product_wishlist is created with the field stock_notification to True
   with a partner associated with the email.
   2. Afterwards, the cron in addons/website_sale_stock/data/ir_cron_data.xml
   will run and send an email to every wishlist email for which the
   stock_notification is True

From a technical point of view, it moves the model product_wishlist from:
website_sale_wishlist -> website_sale.

Since website_sale_stock will use this model to we decided that it was better
to put it there and inherit it in website_sale_wishlist and website_sale_stock

task-id 2806828

+++ addons/website_sale_wishlist/controllers/main.py
@@ -33,15 +33,14 @@ class WebsiteSaleWishlist(WebsiteSale):
             product_id,
             partner_id
         )
-
-        if not partner_id:
-            request.session['wishlist_ids'] = request.session.get('wishlist_ids', []) + [wish.id]
-
+        wish.displayed_in_cart = True
+        request.session.setdefault('wishlist_ids', []).append(wish.id)
         return wish
 
     @route(['/shop/wishlist'], type='http', auth="public", website=True, sitemap=False)
     def get_wishlist(self, count=False, **kw):
         values = request.env['product.wishlist'].with_context(display_default_code=False).current()
+        values = values.filtered(lambda wishlist: wishlist.displayed_in_cart)
         if count:
             return request.make_response(json.dumps(values.mapped('product_id').ids))
 
@@ -59,5 +58,8 @@ class WebsiteSaleWishlist(WebsiteSale):
                 request.session.modified = True
                 wish.sudo().unlink()
         else:
-            wish.unlink()
+            if wish.stock_notification:
+                wish.displayed_in_cart = False
+            else:
+                wish.unlink()
         return True
